fix: refactoring the code to make it run
