+++ .env.example
@@ -6,3 +6,4 @@ GITHUB_API_PROXY=
 NEXT_PUBLIC_REFRESH_INTERVAL_MS=60000
 MOCK_COINGECKO_API=false
 FIRST_AUCTION_BLOCK=0
+DATABASE_URL="file:./dev.db"
