+++ src/app/api/indexer/auction-end-events/routes.ts
@@ -1,13 +1,15 @@
-import { NextApiRequest, NextApiResponse } from 'next';
-import { prisma } from '@/lib/prisma';
-import { rpcFetchBiddingHistory } from '@/queries/rpcFetchBiddingHistory';
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/burn-auction-dapp/blob/main/LICENSE)
+
+import { prisma } from '@/utilities/prisma';
+import { rpcFetchAuctionEnd } from '@/queries/rpcFetchAuctionEnd';
 import { viemPublicClient } from '@/utilities/viem';
 
 const FIRST_AUCTION_BLOCK = process.env.FIRST_AUCTION_BLOCK ? BigInt(process.env.FIRST_AUCTION_BLOCK) : BigInt(0);
 
-export default async function handler(req: NextApiRequest, res: NextApiResponse) {
+export async function GET() {
   try {
-    const latestEvent = await prisma.bidEvent.findFirst({
+    const latestEvent = await prisma.auctionEndEvent.findFirst({
       orderBy: {
         blockNumber: 'desc',
       },
@@ -22,14 +24,15 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
         toBlock = latestBlock;
       }
 
-      const bidEvents = await rpcFetchBiddingHistory(fromBlock, toBlock);
+      const auctionEndEvents = await rpcFetchAuctionEnd(fromBlock, toBlock);
 
-      for (const event of bidEvents) {
-        await prisma.bidEvent.create({
+      for (const event of auctionEndEvents) {
+        await prisma.auctionEndEvent.create({
           data: {
-            sender: event.args.sender,
+            winner: event.args.winner,
             round: event.args.round,
-            amount: event.args.amount,
+            // coins: null,
+            burned: event.args.burned,
             blockNumber: event.blockNumber,
             transactionHash: event.transactionHash,
             transactionIndex: event.transactionIndex,
@@ -44,8 +47,8 @@ export default async function handler(req: NextApiRequest, res: NextApiResponse)
       toBlock = fromBlock + BigInt(10000);
     }
 
-    res.status(200).json({ message: 'Bid events indexed successfully' });
+    return Response.json({ message: 'Auction end events indexed successfully' }, { status: 200 });
   } catch (error) {
-    res.status(500).json({ error: 'Failed to index bid events' });
+    return Response.json({ error: 'Failed to index auction end events' }, { status: 500 });
   }
 }
