+++ src/queries/fetchAuctionHistory.ts
@@ -1,3 +1,6 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/burn-auction-dapp/blob/main/LICENSE)
+
 import { prismaFetchAuctionEvent } from './prismaFetchAuctionEvent';
 import { E } from '@/utilities/error-handling';
 import { Log } from '@/utilities/logger';
