+++ src/queries/prismaFetchAuctionEvent.ts
@@ -1,4 +1,7 @@
-import { prisma } from '@/lib/prisma';
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/burn-auction-dapp/blob/main/LICENSE)
+
+import { prisma } from '@/utilities/prisma';
 
 export const prismaFetchAuctionEvent = async () => {
   try {
