+++ src/queries/prismaFetchBidEvent.ts
@@ -1,10 +1,16 @@
-import { prisma } from '@/lib/prisma';
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/burn-auction-dapp/blob/main/LICENSE)
 
-export const prismaFetchBidEvent = async () => {
+import { prisma } from '@/utilities/prisma';
+
+export const prismaFetchBidEvent = async (round: bigint) => {
   try {
     const bidEvents = await prisma.bidEvent.findMany({
       orderBy: {
         blockNumber: 'desc',
+      },
+      where: {
+        round: round,
       },
     });
