+++ src/queries/rpcFetchBiddingHistory.ts
@@ -1,3 +1,6 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/burn-auction-dapp/blob/main/LICENSE)
+
 import type { BidEvent } from '@/types/BidEvent';
 import { viemPublicClient } from '@/utilities/viem';
 import { CONTRACT_ADDRESS, CONTRACT_ABI } from '@/contract';
