+++ src/utilities/prisma.ts
@@ -1,0 +1,6 @@
+// Copyright Tharsis Labs Ltd.(Evmos)
+// SPDX-License-Identifier:ENCL-1.0(https://github.com/evmos/burn-auction-dapp/blob/main/LICENSE)
+
+import { PrismaClient } from '@prisma/client';
+
+export const prisma = new PrismaClient();
